# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Team AI Collector API
  version: 1.0.0
paths:
  /v1/health:
    get:
      summary: Liveness
  /v1/auth/device:
    post:
      summary: Exchange org API key for device token
  /v1/sync:
    post:
      summary: Ingest sync payload
      description: >
        Accepts an optional top-level `account` string (one of vibe2, vibe3, info, vibe4,
        vibe5) which is stored on the user. Per-session cost (`estimated_cost`) is derived
        server-side from token totals and the session's dominant model.
  /v1/team/overview:
    get:
      summary: Team KPIs for date range
      description: >
        Returns existing KPIs plus cost aggregates (`total_cost`, `today_cost`, `week_cost`,
        `month_cost`, `selected_range_cost`) and `model_distribution`
        (`[{ model, label, count, pct }]`) with `total_usage_count` and `total_token_count`.
      parameters:
        - { name: from, in: query, schema: { type: integer }, description: epoch ms (default now-7d) }
        - { name: to, in: query, schema: { type: integer }, description: epoch ms (default now) }
        - { name: account, in: query, schema: { type: string }, description: optional account filter }
  /v1/team/members:
    get:
      summary: Per-member rollups
      description: >
        Each member now includes `account`, `estimated_cost` (selected range), `total_cost`
        (all-time), `today_cost`, and per-member `model_distribution`.
      parameters:
        - { name: from, in: query, schema: { type: integer } }
        - { name: to, in: query, schema: { type: integer } }
        - { name: account, in: query, schema: { type: string }, description: optional account filter }
  /v1/team/sessions:
    get:
      summary: Paginated session list
      description: >
        Sessions now include `estimated_cost`, `cost_currency`, and `cost_model`.
      parameters:
        - { name: from, in: query, schema: { type: integer } }
        - { name: to, in: query, schema: { type: integer } }
        - { name: user_id, in: query, schema: { type: string } }
        - { name: account, in: query, schema: { type: string }, description: optional account filter }
        - { name: limit, in: query, schema: { type: integer } }
        - { name: offset, in: query, schema: { type: integer } }
  /v1/team/sessions/{id}:
    get:
      summary: Session detail
      description: Returns the full session_stats row, including `estimated_cost`, `cost_currency`, `cost_model`.